refactor: 이미지 뷰어 carousel 인덱스 로직 훅 분리
